Major Refactoring
-Completely refactored MainView, removing its widget-creation capabilities.
-Created TabView parent class for tabs to be based on. TabView is in charge of widget creation, essentially making each tab responsible for its own widgets to better reinforce separation of concerns.
-Created main.py to server as new driver file, again reinforcing separation of concerns by no longer making MainView.py the driver file.
-Refactored GenerationTabView.py to handle its own widget creation.
-Created ResourceManager.py to house resources used by other classes.
-Left GraderTabView.py untouched, still requires refactoring for it to handle its own widget creation.
