+++ AlgorithmView.py
@@ -1,6 +1,7 @@
 from abc import ABC, abstractmethod 
 
 class AlgorithmView(ABC):
+
     def __init__(self, main_view, full_name):
         self.mv = main_view
         self.variables = {}
