+++ main.py
@@ -1,0 +1,6 @@
+from ResourceManager import ResourceManager
+from MainView import MainView
+
+
+resource = ResourceManager()
+view = MainView()
